Revert "Merge branch 'Migration' of https://github.com/DShcharbakou/BE-Oct2021-Exadel-Internship-JS-.NET-BA into Migration"
This reverts commit f7b1a8ca3749afef445b996bda114324ed953ded, reversing
changes made to 1f3d9148bd9922559ecabae9e97520186ed9ffec.

+++ DAL/Repositories/BaseRepository.cs
@@ -24,7 +24,7 @@ namespace DAL.Repositories
             return _dbSet.ToList();
         }
 
-        public virtual T Get(int id)
+        public virtual T Get(long id)
         {
             return _dbSet.SingleOrDefault(x => x.Id == id);
         }
@@ -46,7 +46,7 @@ namespace DAL.Repositories
             _internshipDbContext.Remove(model);
         }
 
-        public virtual void Remove(int id)
+        public virtual void Remove(long id)
         {
             var model = Get(id);
             Remove(model);
